Move WelcomeScreen widget to separate file
