+++ test/auth_section/welcome_screen/widget_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:niira/main.dart';
 import 'package:niira/screens/create_account.dart';
 import 'package:niira/screens/sign_in.dart';
+import 'package:niira/screens/welcome.dart';
 
 class MockNavigatorObserver extends Mock implements NavigatorObserver {}
